This PR is mainly focused on providing a common set of utility functions and translator for reading and translating UsdGeomMesh prim into a Maya mesh object.To make this happen I had to make some architectural changes in UsdMayaTranslatorMesh class:
-  Create a new class interfaces with no internal dependency to UsdMayaPrimReaderArgs and UsdMayaPrimReaderContext.

- Remove private static functions, extract their operations and provide as a new set of utilities under meshUtils ( e.g assignPrimvarsToMesh, assignInvisibleFaces, assignSubDivTagsToMesh ).

- Provide getter functions to get access Maya Mesh object as well as deforming meshes (e.g blendshape, PointBasedDeformer) objects.
