+++ lib/usd/translators/meshReader.cpp
@@ -15,17 +15,42 @@
 //
 // Modifications copyright (C) 2020 Autodesk
 //
-
 #include "pxr/pxr.h"
 
 #include "../../fileio/primReaderRegistry.h"
+#include "../../fileio/translators/translatorGprim.h"
+#include "../../fileio/translators/translatorMaterial.h"
 #include "../../fileio/translators/translatorMesh.h"
+#include "../../fileio/translators/translatorUtil.h"
+#include "../../fileio/utils/meshUtil.h"
+#include "../../fileio/utils/readUtil.h"
+#include "../../nodes/stageNode.h"
+
+#include <maya/MFnBlendShapeDeformer.h>
 
 #include "pxr/usd/usdGeom/mesh.h"
 
 PXR_NAMESPACE_OPEN_SCOPE
 
-// Prim reader for mesh
+namespace
+{
+    bool
+    assignMaterial(const UsdGeomMesh& mesh,
+                   const UsdMayaPrimReaderArgs& args,
+                   const MObject& meshObj,
+                   UsdMayaPrimReaderContext* context)
+    {
+        // If a material is bound, create (or reuse if already present) and assign it
+        // If no binding is present, assign the mesh to the default shader
+        const TfToken& shadingMode = args.GetShadingMode();
+        return UsdMayaTranslatorMaterial::AssignMaterial(shadingMode,
+                                                         mesh,
+                                                         meshObj,
+                                                         context);
+    }
+}
+
+// prim reader for mesh
 class MayaUsdPrimReaderMesh final : public UsdMayaPrimReader
 {
 public:
@@ -48,14 +73,90 @@ TF_REGISTRY_FUNCTION_WITH_TAG(UsdMayaPrimReaderRegistry, UsdGeomMesh)
 bool
 MayaUsdPrimReaderMesh::Read(UsdMayaPrimReaderContext* context)
 {
-    const UsdPrim& usdPrim = _GetArgs().GetUsdPrim();
+    if(!context){
+        return false;
+    }
 
-    auto parentNode = context->GetMayaNode(usdPrim.GetPath().GetParentPath(), true);
+    MStatus status;
 
-    return UsdMayaTranslatorMesh::Create(UsdGeomMesh(usdPrim),
-                                         parentNode,
-                                         _GetArgs(),
-                                         context);
+    const auto& prim = _GetArgs().GetUsdPrim();
+    auto mesh = UsdGeomMesh(prim);
+    if (!mesh) {
+        return false;
+    }
+
+    auto parentNode = context->GetMayaNode(prim.GetPath().GetParentPath(), true);
+    MObject transformObj;
+    bool retStatus = UsdMayaTranslatorUtil::CreateTransformNode(prim,
+                                                                parentNode,
+                                                                _GetArgs(),
+                                                                context,
+                                                                &status,
+                                                                &transformObj);
+    if(!retStatus){
+        return false;
+    }
+
+    // get the USD stage node from the context's registry
+    MObject stageNode;
+    if (_GetArgs().GetUseAsAnimationCache()){
+        stageNode = context->GetMayaNode(SdfPath(UsdMayaStageNodeTokens->MayaTypeName.GetString()),false);
+    }
+
+    MayaUsd::TranslatorMeshRead meshRead(mesh, 
+                                         prim, 
+                                         transformObj, 
+                                         stageNode, 
+                                         _GetArgs().GetTimeInterval(),
+                                         _GetArgs().GetUseAsAnimationCache(),
+                                         &status);
+    CHECK_MSTATUS_AND_RETURN(status, false);
+
+    // shape name
+    const auto& primName = prim.GetName().GetString();
+    status = UsdMayaMeshUtil::setName(meshRead.meshObject(), primName);
+    CHECK_MSTATUS_AND_RETURN(status, false);
+
+    // mesh is a shape, so read Gprim properties
+    UsdMayaTranslatorGprim::Read(mesh, meshRead.meshObject(), context);
+
+    // undo/redo mesh object
+    const auto shapeName = TfStringPrintf("%sShape", primName.c_str());
+    const auto shapePath = prim.GetPath().AppendChild(TfToken(shapeName));
+    context->RegisterNewMayaNode(shapePath.GetString(), meshRead.meshObject());
+
+    // undo/redo deformable mesh (blenshape, PointBasedDeformer)
+    if (meshRead.pointsNumTimeSamples() > 0) {
+        if (_GetArgs().GetUseAsAnimationCache()){
+            context->RegisterNewMayaNode(meshRead.pointBasedDeformerName().asChar(), 
+                                         meshRead.pointBasedDeformerNode());
+        }
+        else {
+            if(meshRead.blendObject().apiType() == MFn::kBlend){ 
+                return false;
+            }
+
+            MFnBlendShapeDeformer blendFnSet(meshRead.blendObject());
+            context->RegisterNewMayaNode(blendFnSet.name().asChar(), 
+                                         meshRead.blendObject());
+        }
+    }
+
+    // assign material
+    assignMaterial(mesh, _GetArgs(), meshRead.meshObject(), context);
+
+    // assign primvars to mesh
+    UsdMayaMeshUtil::assignPrimvarsToMesh( mesh, 
+                                           meshRead.meshObject(),
+                                           _GetArgs().GetExcludePrimvarNames());
+    // assign invisible faces
+    UsdMayaMeshUtil::assignInvisibleFaces(mesh, meshRead.meshObject());
+
+    if(!retStatus){
+        return false;
+    }
+
+    return true;
 }
 
 PXR_NAMESPACE_CLOSE_SCOPE
